backend/requirements.txt
earthengine-api
fastapi
pydantic
google-auth
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
